+++ package-kuring/Sources/UIKit/NoticeUI/Resources/Symbols.xcassets/navigation/bell.imageset/Contents.json
@@ -1,16 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "bell.svg",
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
       "filename" : "bell-dark.svg",
       "idiom" : "universal"
     }
